% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse-funs.R
\name{create_multi}
\alias{create_multi}
\title{Create the scenarios for the multiverse}
\usage{
create_multi(formula, focal = NULL, nfuns = NULL, cfuns = NULL, data)
}
\arguments{
\item{formula}{a formula describing the maximal model with bare variables (no tranformations). Currently not supporting interactions.}

\item{focal}{optional name of the coefficient that is the focus of the analysis. No transformation will be applied to that variable and no model will exclude this variable.}

\item{nfuns}{functions to be applied to numerical variables. functions need to be provided as characters. this argument can be a vector of functions (e.g., \code{c("log", "exp")}) and the functions will be applied al all numerical variable exluding the focal predictor. In alternative, can be a named list specifing the name of the variable and the functions as string (e.g., \code{list(x = "log", z = c("exp"))}) in this way the functions are variable-specific. If \code{NULL} no transformations will be applied.}

\item{cfuns}{same as the \code{nfuns} but for factor/character variables. If \code{NULL} no transformations will be applied.}

\item{data}{the dataset for the model}
}
\value{
a list
}
\description{
Create the scenarios for the multiverse
}
